% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interperscon.R
\name{interpers_con}
\alias{interpers_con}
\title{Interpersonal Consistency}
\usage{
interpers_con(data, id)
}
\arguments{
\item{data}{yourdataset}

\item{id}{youridvariable}
}
\description{
Calculates the extent of interpersonal consistency for the adjective ranking task described by Block (1961). Instead of the centroid method of factor analysis, which was used in the paper, an eigen decomposition is done. "The "data" needs to be in long format with "id" as key variable. The package returns one value for each particpant, which is the extent of interpersonal consistency. A higher value means a higher interpersonal consistency, the value can range from 0 to 100.
}
\examples{
interpers_con(yourdata, youridvariable)

}
\references{
Block, J. (1961). Ego Identity, Role Variability, and Adjustment. Journal of Clinical Psychology, 25(5), 392-397.
}
\author{
Esther Beierl <esther.beierl@gmx.de> [‘aut‘, ‘cre‘], Lawrence Yu [‘ctb‘]
}
